Doc: Tests fixes
